fix: footer link should not spoil the gameplay
On touch devices, we want to see the game, so we mainly play we our finger under, on the footer (it also more accessible).
But here is the signature with the link to the portfolio...
And when the game is on, a touch should pause the game, and not spoil the xp by opening a link.

- So we need the disable the link on touch when the game is playing.
- And more and more device can have a mouse AND a touch screen.
So we can't rely on the simple fact that there is a touch screen to disable the link
But on many touch device, the click event is simulated, behaved like a mouseEvent with no way to know if it was triggered by touch.
So we need to listened to touch event before.
- Need to cancel this disactivation if the click is not trigger after (other touch move), for user with touch screen AND a mouse